kitty: fix issue with serial input
The problem with always waiting on a event to say that we
have new serial input is that sometimes there isn't anything
new, although we have more data left to process.

The particular scenario is when MicroPython has not been
executed for a long time although we have had serial input
occur. This would have resulted in multiple new entries in
the shared queues, but only a single notification.