Update to v1.22.2 of MicroPython
At the time of writing, this is the latest release of MicroPython.

The main motivation for upgrading is that we were observing very slow
imports of scripts over the REPL via NFS. This was because MicroPython
was limited to reading 24 bytes at a time. This lead to small scripts
taking five or more seconds to load.

The commit 4cf741062b20fa34062d236d56b2168515e990b4 fixes this by
both increasing the maximum buffer size and allowing VFS implementations
to return an ideal buffer size. It is still not completely ideal as
the maximum buffer is still fairly low at 255 bytes.