nfs: fix POSIX socket emulation layer
The NFS client did not reconnect to the NFS server when its connection
was dropped. There were two issues here, which this commit fixes:
first, the NFS client did not attempt to reconnect when its connection
was dropped; and second, if it were to try to reconnect, it would
fail. This commit fixes both.

The first issue was not a matter of implementing reconnect
functionality, as libnfs already provides this, but rather the issue
was that we were not properly notifying libnfs that the connection had
dropped by setting the POLLHUP bit in the events parameter passed into
nfs_service. We now correctly differentiate between a closed
connection due to libnfs calling close() and a closed connection due
to other factors, and report POLLHUP in the latter case.

The second issue was that even if libnfs were to try to reconnect, it
would fail due to its usage of dup2 on its socket file desctiptors
when reconnecting. Previous to this commit, dup2 on socket FDs was not
implemented correctly in the POSIX emulation layer. This commit
re-architectures the socket code to fix this.