wip: kitty communication with server
Made some progress on getting Kitty to communicate with the python
server. Still quite a bit of instability. The main issue is that
the connection gets reset every once in a while because the server
does not receive a heartbeat from the kitty in time. Lowering the
time between heartbeats helps, but does not solve the problem. We
have investigated the micropython side of it a bit, and it seems like
the co-routine for sending the heartbeat is called on time, so the
problem lies either in the implementations of write() or drain(), or
server-side. What was observed from the server is that there are times
when no heartbeats appear for several seconds, followed by several at
once.