kitty: minor fixes
Ivan's previous change to modnetwork made the communication between the
kitty client and server much more reliable. It still isn't perfect, as
there is occasionally a slight (but noticeable) delay between when a
packet is queued up to be sent, and when it is actually sent/received
by the server. I believe this is because the notify() only occurs in
the event loop, but the netif_output function does not co_switch,
meaning that the packet will only be sent when something else in
the system triggers a co_switch back to the event loop.