kitty: Got pypng to be imported
James' kitty client code depends on libpng, which this commit makes
importable. It required the addition of the itertools, operator, zlib,
and png libraries to the NFS as well as the addition of modre and
moddeflate in the config. This has not been properly tested, and the
micropython heap seems to currently be too small to import the specified
image.