kitty: deal with libm.a better
MicroPython needs to link libm in order to be able to use certain
functionality. Now, we can't just give it the path to libm.a in
the toolchain since that also has libc.a and we are already compiling
MicroPython with a specific libc.a and so we'll get a clash.

This is a terrible hack, and needs to be removed as soon as possible
by switching over to everything using a single libc.