mp: move micropython port code to components/
This commit moves the MicroPython port directory from
examples/kitty/src/micropython to components/micropython. The code in
this directory currently builds to a proper ELF file which can be run
as a Microkit PD. Currently it will always build a MP program that
runs a REPL, and always has I2C and framebuffer enabled, among other
things. In the future we may either want to make these things
configurable.

Signed-off-by: James Archer <j.archer@unsw.edu.au>