nfs: set timeout in correct location
If we set the timeout unconditionally in the notified function, then
notifications from channels other than the timer will cause the
timeout to be re-set, delaying the timeout. This can cause the code
that is run upon receiving a timeout to be run much later than it
should have been (or theoretically never).

Signed-off-by: James Archer <j.archer@unsw.edu.au>