mp: enable additional time module functionality
This enables extra functionality in the time module such as the
gmtime function which is required by the webserver system. However, it
requires an implementation of localtime. We cannot provide a real
implementation due to a lack of real-time clock on the Odroid-C4
system, so instead the function raises a not implemented exception.

Signed-off-by: James Archer <j.archer@unsw.edu.au>