fs: change fstat to fsize
While on a Unix system, the VFS layer will implement fstat, the
LionsOS filesystem protocol poses requirements on filesystem
components directly, and fstat may not be something we want
filesystems to have to implement. Generally, the full stat operation
will be an operation on a directory rather than a file, so having a
stat operation on a file descriptor will require the filesystem to
maintain associations between directory and open files. This is a
requirement we may not want to impose.

However, implementing Python's open command requires implementing
append, and since we do not implement non-positional read/write,
append must be implemented client-side. That then requires us to get
the size of a file while opening it, so that we can set the user-side
file pointer to to the size of the file, so that writes append to the
file.

It is possible to do this on a local filesystem simply by stating and
then opening, provided we have the requirement that the same file
cannot be opened twice, but on a network filesystem like NFS, we have
the well-known race condition between stat and open, in that the size
of the file you have opened may not be the size of the file on disk
read by stat, as another process could have opened and written to the
file in the meantime. Thus on Unix there is a need to get the size of
the *opened* file, hence fstat. But given that we need only the size
and not the rest of the data, we can loosen this to fsize.

Signed-off-by: James Archer <j.archer@unsw.edu.au>
Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>