ci: build examples on Ubuntu using apt package manager
All of our CI currently uses Nix to fetch packages since it has
been more reliable but this means that users who use apt sometimes
run into issues that our CI has not.

By trying to emulate a popular developer environment in CI, we can
hopefully catch and fix these issues before our users do.

One problem is that in our examples we use GCC since we depend on
newlibc and had run-time issues when building libnfs with LLVM.
This meant that in some cases we are mixing LLVM built and GCC
built objects which does not seem to work on older versions of LLD.
Because of this, we update the Ubuntu version to 24.04 which packages
newer versions of LLVM tools. This does not solve the root problem,
which requires more investigation and work.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>