nfs: fix tcp socket leak
Our socket structure wraps the lwip socket PCB object. When we free
one of our sockets, we close the lwip PCB and set our socket to
unallocated. However, closing an lwip PCB doesn't always immediately
end that object's lifetime: sometimes it can sit around for a while
due to how closing a socket works in TCP. If this happens and then the
closed PCB enters an error state, lwip will call our socket error
callback. Previously, the socket error callback would set the socket
structure that tracked the lwip PCB to an error state. However, in
such a case, the socket structure was in an unallocated state, having
already been closer prior. This meant that the socket leaked and could
never be used again.

This commit changes the socket close code to also unset the lwip PCB
callback argument, so that if this happens and lwip calls our socket
error callback function on a PCB that once belonged to a socket that
has already been deallocated, we won't change the socket's state to
error but will instead leave it in an unallocated state.

Signed-off-by: James Archer <j.archer@unsw.edu.au>