fat: remove wrapper around libmicrokitco
The wrapper around libmicrokitco, defined in co_helper.c/h, adds lines
of code rather than reducing them, and having a layer of indirection
makes reading the code harder. Beyond that, libmicrokitco is
first-party code and if the interface is lacking, we should improve it
at the source rather than wrapping it in our other code.

Signed-off-by: James Archer <j.archer@unsw.edu.au>