fat: add FAT filesystem component
This commit adds a new LionsOS component 'fat' which implements the LionsOS
filesystem protocol. The fat component uses the sDDF-blk protocol to
act as a filesystem for a single FAT/exFAT partition on a block
device. It depends on Elm-chan's FatFs library, which this commit adds
as an in-tree dependency (dep/ff15).

Signed-off-by: James Archer <j.archer@unsw.edu.au>
Co-authored-by: Cheng Li <lichengchaoreng@gmail.com>