Fix subnet mask calculation
The subnet mask was calculated as if bit ordering were reversed
inside each byte, as well as in the whole word.  (i.e., it treated
1<<n little endian applied to a big endian word as if it were
(0x80000000>>n) for a litte endian word.)  This only works if n is a
mutliple of 8.

Change to something that works all the time, and remove now unnecessary
special-casing for a 32-bit mask.

Signed-off-by: Peter Chubb <Peter.Chubb@unsw.edu.au>