micropython: update to v1.26.1
The only reason for the update is that Clang 21 seems to have
an extra warning that ends up as an error which means LionsOS
examples don't build unless you have Clang 20 or below.

Upstream has fixed this already, so we just have to update.

USER_C_MODULES is slightly different semantics to want we want
right now, so change the name to not clash with what MicroPython
expects.

Co-authored-by: Peter Chubb <Peter.Chubb@unsw.edu.au>
Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>